Implemented Database.table_constructor which can be used to dictate an AdvancedTable Subclass to be used to initiliaze AdvancedTables on a per-table basis.
