+++ alcustoms.pyproj
@@ -131,6 +131,9 @@
     </Compile>
     <Compile Include="sql\objects\View.py" />
     <Compile Include="sql\objects\__init__.py" />
+    <Compile Include="sql\tests\objects\test_Connection.py">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="sql\tests\objects\test_graphdb.py" />
     <Compile Include="sql\tests\objects\test_Table.py">
       <SubType>Code</SubType>
